refactor: 불필요한 파일 제거 및 파일 이동
